+++ LibBrotli/src/decode.jl
@@ -47,9 +47,7 @@ function try_decode!(d::BrotliDecodeOptions, dst::AbstractVector{UInt8}, src::Ab
 end
 
 function try_resize_decode!(d::BrotliDecodeOptions, dst::AbstractVector{UInt8}, src::AbstractVector{UInt8}, max_size::Int64; kwargs...)::Union{Nothing, Int64}
-    check_in_range(Int64(0):max_size; dst_size=length(dst))
-    olb::Int64 = length(dst)
-    dst_size::Int64 = olb
+    dst_size::Int64 = length(dst)
     src_size::Int64 = length(src)
     src_left::Int64 = src_size
     dst_left::Int64 = dst_size
@@ -96,17 +94,7 @@ function try_resize_decode!(d::BrotliDecodeOptions, dst::AbstractVector{UInt8}, 
                 end
                 if result == BROTLI_DECODER_RESULT_NEEDS_MORE_OUTPUT
                     @assert iszero(dst_left)
-                    # grow dst or return nothing
-                    if dst_size ≥ max_size
-                        return nothing
-                    end
-                    # This inequality prevents overflow
-                    local next_size = if max_size - dst_size ≤ dst_size
-                        max_size
-                    else
-                        max(2*dst_size, Int64(1))
-                    end
-                    resize!(dst, next_size)
+                    local next_size = @something grow_dst!(dst, max_size) return nothing
                     dst_left += next_size - dst_size
                     dst_size = next_size
                     @assert dst_left > 0
@@ -115,9 +103,6 @@ function try_resize_decode!(d::BrotliDecodeOptions, dst::AbstractVector{UInt8}, 
                         # yay done return decompressed size
                         real_dst_size = dst_size - dst_left
                         @assert real_dst_size ∈ 0:length(dst)
-                        if length(dst) > olb # shrink to just contain output if it was resized.
-                            resize!(dst, real_dst_size)
-                        end
                         return real_dst_size
                     else
                         # Otherwise, throw an error